requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis for caching and sessions
redis==5.0.1
aioredis==2.0.1

# Authentication and Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Background tasks and scheduling
celery==5.3.4
kombu==5.3.4

# File handling and storage
boto3==1.34.0
Pillow==10.1.0

# Email
fastapi-mail==1.4.1

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Geolocation
geopy==2.4.1

# Date and time
python-dateutil==2.8.2

# Validation
email-validator==2.1.0

# PDF generation
reportlab==4.0.7

# Excel handling
openpyxl==3.1.2

# HTTP client
httpx==0.25.2

# Rate limiting
slowapi==0.1.9

# WebSocket support
websockets==12.0

# Configuration
python-dotenv==1.0.0

# Biometric integration
requests==2.31.0

# Multi-language support
babel==2.13.1

# Currency conversion
forex-python==1.8
